Run --pdb check early
Run check for --pdb as early as possible so it kicks in _before_ plugin loading and other dynamic startup happens
