+++ .flake8
@@ -6,6 +6,9 @@ max-line-length = 80
 max-complexity = 18
 select = B,C,E,F,W,T4,B9
 
+per-file-ignores =
+    nixops/__main__.py:E402
+
 # # We need to configure the mypy.ini because the flake8-mypy's default
 # # options don't properly override it, so if we don't specify it we get
 # # half of the config from mypy.ini and half from flake8-mypy.
